+++ scripts/ci/configure_debug_fallback.sh
@@ -13,5 +13,4 @@ meson \
   --prefix "${PREFIX_ARG:-/usr/local}" \
   --buildtype=debugoptimized \
   -Dtests="${ENABLED_TESTS:-false}" \
-  -Dsse41="${ENABLED_SSE41:-false}" \
   "${BUILD_DIR}" .